pyproject.toml
[project]
name = "nexly-agent-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "mistralai>=1.12.4",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "prophet>=1.3.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
